Move hooks-related classes to their own package
