+++ src/main/java/com/greghaskins/spectrum/internal/Parent.java
@@ -1,5 +1,7 @@
 package com.greghaskins.spectrum.internal;
 
+import com.greghaskins.spectrum.internal.hooks.Hooks;
+
 public interface Parent {
 
   void focus(Child child);
