+++ src/main/java/com/greghaskins/spectrum/internal/hooks/AbstractSupplyingHook.java
@@ -1,4 +1,4 @@
-package com.greghaskins.spectrum.internal;
+package com.greghaskins.spectrum.internal.hooks;
 
 import static com.greghaskins.spectrum.Spectrum.assertSpectrumInTestMode;
