+++ src/main/java/com/greghaskins/spectrum/internal/hooks/Hook.java
@@ -1,4 +1,4 @@
-package com.greghaskins.spectrum.internal;
+package com.greghaskins.spectrum.internal.hooks;
 
 import com.greghaskins.spectrum.Block;
 import com.greghaskins.spectrum.ThrowingConsumer;
